[Refactor] Move Iceberg sink out of Exec
