+++ be/src/exec/data_sinks/data_sink_pipeline.cpp
@@ -52,7 +52,7 @@
 #include "gen_cpp/Exprs_types.h"
 #include "gen_cpp/Partitions_types.h"
 #ifndef __APPLE__
-#include "exec/data_sinks/iceberg_table_sink.h"
+#include "data_sink/external/iceberg_table_sink.h"
 #include "exec/data_sinks/iceberg_table_sink_pipeline_builder.h"
 #endif
 #include "data_sink/exchange/multi_cast_data_stream_sink.h"
