+++ be/test/connector_sink/iceberg_table_sink_test.cpp
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "exec/data_sinks/iceberg_table_sink.h"
+#include "data_sink/external/iceberg_table_sink.h"
 
 #include <gmock/gmock.h>
 #include <gtest/gtest-param-test.h>
